Première version du menu
La première version du menu console du jeu a les fonctionnalités
suivantes :
- Choisir de jouer une nouvelle partie
- Charger une partie déjà existante
- Quitter l'application
