+++ Sources/menu.cpp
@@ -1,17 +1,9 @@
 #include <iostream>
 #include <iomanip>
 #include <cstdlib>
-#include "menu.h"
+#include "../Headers/menu.h"
 
 using namespace std;
-
-// Definition des methodes de la classe OptionMenu
-
-OptionMenu::OptionMenu(const string &nom, const string &description)
-: nom_(nom), description_(description)
-{
-}
-
 
 // Definition des methodes de la classe Menu
 
@@ -79,6 +71,11 @@ bool Menu::quitter()
 	char reponse;
 	cout << "Voulez-vous vraiment sortir de l'application (o/n) ? ";
 	cin >> reponse;
+	while ((reponse != 'o') || (reponse != 'O') || (reponse != 'n') || (reponse != 'N'))
+	{
+			cout << "Reponse erronnee. Veuillez ecrire 'o' ou 'n' : ";
+			cin >> reponse;
+	}
 	if ((reponse == 'o') || (reponse == 'O')) fin = true;
 	return fin;
 }
