+++ Sources/optionMenu.cpp
@@ -1,0 +1,7 @@
+#include "../Headers/optionMenu.h"
+
+// Definition des methodes de la classe OptionMenu
+OptionMenu::OptionMenu(const string &nom, const string &description)
+		: nom_(nom), description_(description)
+{
+}
